added public invite link component
