+++ app/src/components/user/AccountIndexPage/ManageAccount/ManageTeams/TeamViewer/MembersDetails/index.js
@@ -1,8 +1,8 @@
-import React, { useState, useEffect, useCallback } from "react";
+import React, { useState, useEffect } from "react";
 import { useSelector } from "react-redux";
 import { useLocation } from "react-router-dom";
-import { Col, Row, Button, Typography, Input, Divider } from "antd";
-import { getAvailableTeams } from "store/features/teams/selectors";
+import { Col, Row, Button, Typography } from "antd";
+import { getAvailableTeams, getCurrentlyActiveWorkspaceMembers } from "store/features/teams/selectors";
 //Firebase
 import TeamMembersTable from "./TeamMembersTable";
 import { getFunctions, httpsCallable } from "firebase/functions";
@@ -11,28 +11,30 @@ import { PlusOutlined } from "@ant-design/icons";
 // Sub Components
 import AddMemberModal from "./AddMemberModal";
 import { trackAddMemberClicked } from "modules/analytics/events/common/teams";
-import LearnMoreAboutWorkspace from "../common/LearnMoreAboutWorkspace";
 import "./MembersDetails.css";
-import { RQButton } from "lib/design-system/components";
-import CopyButton from "components/misc/CopyButton";
-import { toast } from "utils/Toast";
+import { getUserAuthDetails } from "store/selectors";
+import PublicInviteLink from "./PublicInviteLink";
 
 const MembersDetails = ({ teamId }) => {
   const location = useLocation();
   const isNewTeam = location.state?.isNewTeam;
+
   // Component state
   const [seats, setSeats] = useState({});
   const [showSeatStatus, setShowSeatStatus] = useState(false);
   const [isAddMemberModalActive, setIsAddMemberModalActive] = useState(false);
   const [refreshTeamMembersTable, setRefreshTeamMembersTable] = useState(false);
-  const [publicInviteId, setPublicInviteId] = useState(null);
-  const [publicInviteLoading, setPublicInviteLoading] = useState(false);
 
   // Global state
   const availableTeams = useSelector(getAvailableTeams);
+  const currentTeamMembers = useSelector(getCurrentlyActiveWorkspaceMembers);
+  const user = useSelector(getUserAuthDetails);
+  const isCurrentUserAdmin = currentTeamMembers[user?.details?.profile?.uid]?.isAdmin === true;
+
+  console.log("isCurrentUserAdmin", isCurrentUserAdmin);
   const teamDetails = availableTeams?.find((team) => team.id === teamId) ?? {};
   const accessCount = teamDetails?.accessCount;
-
+  
   // To handle refresh in TeamMembersTable
   const doRefreshTeamMembersTable = () => {
     setRefreshTeamMembersTable(!refreshTeamMembersTable);
@@ -73,108 +75,9 @@ const MembersDetails = ({ teamId }) => {
       .catch(() => setShowSeatStatus(false));
   }, [teamId, refreshTeamMembersTable]);
 
-  const handlePublicInviteCreateClicked = () => {
-    const functions = getFunctions();
-    const createTeamInvite = httpsCallable(functions, "invites-createTeamInvite");
-    createTeamInvite({ teamId: teamId, usage: "unlimited"})
-      .then((res) => {
-        if(res?.data?.success) {
-          setPublicInviteId(res?.data?.inviteId);
-        } else {
-          toast.error("Only admins can invite people");
-        }
-      })
-  }
-
-  const handlePublicInviteRevokeClicked = () => {
-    const functions = getFunctions();
-    const revokeInvite = httpsCallable(functions, "invites-revokeInvite");
-    revokeInvite({ inviteId: publicInviteId })
-      .then((res) => {
-        if(res?.data?.success) {
-          setPublicInviteId(null);
-          toast.success("Successfully Revoked invite");
-        } else {
-          toast.error("Only admins can revoke invites");
-        }
-      })
-  }
-
-  const generateInviteLinkFromId = (inviteId) => {
-    return `${window.location.origin}/invite/${inviteId}`;
-  }
-
-  const fetchPublicInviteLink = () => {
-    setPublicInviteLoading(true);
-    const functions = getFunctions();
-    const getTeamPublicInvite = httpsCallable(functions, "invites-getTeamPublicInvite");
-    getTeamPublicInvite({ teamId: teamId })
-      .then((res) => {
-        if(res?.data?.success) {
-          setPublicInviteId(res?.data?.inviteId);
-        }
-        setPublicInviteLoading(false);
-      })
-      .catch(err => {
-        setPublicInviteLoading(false);
-      })
-  }
-
-  const stableFetchPublicInviteLink = useCallback(fetchPublicInviteLink, [teamId]);
-
-  useEffect(() => {
-    stableFetchPublicInviteLink();
-  }, [stableFetchPublicInviteLink]);
-
-
   return (
     <div className="members-details-container">
-      { publicInviteLoading ? null : (
-        <Row>
-          <Col span={24}>
-            <LearnMoreAboutWorkspace
-              linkText="Learn about adding members to your
-              workspace"
-            />
-
-            <Row align="middle" justify="space-between">
-              <Col className="header members-invite-title">Public Invite link</Col>
-              <Col className="ml-auto">
-              {
-                publicInviteId?
-                (<RQButton type="danger" onClick={handlePublicInviteRevokeClicked}>Revoke</RQButton>):
-                (<RQButton onClick={handlePublicInviteCreateClicked} type="primary">Create Link</RQButton>)
-              }
-              </Col>
-            </Row>
-
-            <p className="text-dark-gray members-info-message">
-              Share this secret link to invite people to this workspace. Only
-              users who can invite members can see this.
-            </p>
-          </Col>
-        </Row>
-      ) }
-
-      {publicInviteId? (
-        <Row className={"public-invite-link-container"} justify="space-between">
-          <Col span={24}>
-            <Input
-              className="invite-link-input"
-              contentEditable={false}
-              value={generateInviteLinkFromId(publicInviteId)}
-              addonAfter={<CopyButton
-                title=""
-                copyText={generateInviteLinkFromId(publicInviteId)}
-              />}
-              disabled={true}
-              type="text"
-            />
-          </Col>
-        </Row>
-      ): null}
-
-      <Divider className="manage-workspace-divider" />
+      <PublicInviteLink teamId={teamId} />
 
       {/* members table */}
       <Row>
